+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Use vdpau-va-gl for vdpau software rendering (#390)
 * Implement multiple SDL audio devices
 * Clicking the first column seeks to frame in input editor
+* Lua memory read and write float/double
 
 ### Changed